irb_assistant/requirements.txt
openai
python-dotenv
streamlit
PyPDF2
python-docx
# Add a vector store library, e.g.:
chromadb
# Or:
# faiss-cpu
# sentence-transformers # Often needed for embeddings